wip: Various cliff.*.toml files
- clif.git-cliff.toml is the cliff.toml from
  https://github.com/orhun/git-cliff/

- cliff.rp-rs.github.toml is "close" to what I think the default
  configuration and named Cliff.toml. But see comment in the file.

- cliff.rp-rs.rp-hal.plus-pr_title.toml is a minimalist version that
  prints just version and commit info.

- merge_commit-true.png is the "rendered" version I posted to discord
  decided to keep it for no real good reason.

Note the corresponding files with cliff.winksaville.* files are
identical except the owner is "winksaville" instead of "rp-rs". This
shows that only the "primary" repo (rp-rs/rp-hal) has the pr_url info
and "forked" repos (winksaville/rp-hal) do not. So the owere must point
at the primary repo to create a CHANGELOG.